+++ src/pymetropolis/metro_demand/population/generic.py
@@ -47,11 +47,8 @@ class GenericPopulationStep(PopulationStep):
         self.output["persons"].write(persons)
 
 
-class PopulationFromTripCoordinatesStep(GeoStep, PopulationStep):
-    """Generates a population (persons and trips) from a list of trips with origin / destination
-    coordinates.
-
-    Each person has a single trip.
+class ImportTripCoordinatesStep(GeoStep, PopulationStep):
+    """Imports trips from a file with origin / destination coordinates for each trip.
 
     The input file must have columns: `trip_id`, `origin_lng`, `origin_lat`, `destination_lng`,
     `destination_lat`.
@@ -62,19 +59,13 @@ class PopulationFromTripCoordinatesStep(GeoStep, PopulationStep):
         check_file_exists=True,
         description="Path to a Parquet / CSV file with coordinates of each trip.",
     )
-    output_files = {
-        "trips": TripsFile,
-        "persons": PersonsFile,
-        "origins": TripsOriginsFile,
-        "destinations": TripsDestinationsFile,
-    }
+    output_files = {"origins": TripsOriginsFile, "destinations": TripsDestinationsFile}
 
     def is_defined(self):
         return self.trip_coordinates_file is not None
 
     def run(self):
         import geopandas as gpd
-        import polars as pl
 
         assert self.trip_coordinates_file is not None
 
@@ -96,6 +87,28 @@ class PopulationFromTripCoordinatesStep(GeoStep, PopulationStep):
         destinations_gdf = gpd.GeoDataFrame(
             {"trip_id": df["trip_id"]}, geometry=destinations
         ).to_crs(self.crs)
+        self.output["origins"].write(origins_gdf)
+        self.output["destinations"].write(destinations_gdf)
+
+
+class PopulationFromTripCoordinatesStep(GeoStep, PopulationStep):
+    """Generates a population (persons and trips) from trips' origin / destination.
+
+    Each person has a single trip.
+    """
+
+    input_files = {"origins": TripsOriginsFile, "destinations": TripsDestinationsFile}
+    output_files = {"trips": TripsFile, "persons": PersonsFile}
+    priority = 0
+
+    def run(self):
+        import polars as pl
+
+        # We don't actually need to read the destinations but we still keep them in the input files
+        # to make sure that they exist.
+        origins = self.input["origins"].read()
+        df = pl.from_pandas(origins.loc[:, ["trip_id"]])
+
         trips = df.select(
             "trip_id",
             person_id="trip_id",
@@ -110,8 +123,6 @@ class PopulationFromTripCoordinatesStep(GeoStep, PopulationStep):
             has_driving_license=pl.lit(True, dtype=pl.Boolean),
             has_public_transit_subscription=pl.lit(True, dtype=pl.Boolean),
         )
-        self.output["origins"].write(origins_gdf)
-        self.output["destinations"].write(destinations_gdf)
         self.output["trips"].write(trips)
         self.output["persons"].write(persons)